Merge feature/2.2.21-llvm19.1.4 into develop-2.2.21-OH: Add CRT allocator and CMC GC support

Created-by: liuhanzhao
Commit-by: liuhanzhao;mafanxii;MRXVAM;chuxu6;kaitianlu;Igor Iorkh;cc103546;Xiaowenhu96
Merged-by: QuyoueryuanDCar
Description: 
> **本模板适用于合并到 `develop-*` 分支的日常变更。**  
> 如果目标是 `next/*` 分支，请使用 `next.md` 模板；如果目标是 `main-*` 分支，请使用 `release.md` 模板。

> 💡 **首次提交 Pull Request？**  
> 请先阅读贡献指南：[CONTRIBUTING.md](https://gitcode.com/OpenHarmony-CrossPlatformFramework/community/blob/main/CONTRIBUTING.md)

## 变更类型
- [ ] 新功能
- [ ] Bug 修复
- [ ] 性能优化
- [ ] 代码重构
- [ ] 文档更新
- [ ] 测试补充

## 变更描述
<!-- 简要说明本次修改的内容和目的 -->

## 测试情况
<!-- 请说明测试覆盖情况，包括单元测试、集成测试等 -->

## 适配验证
- [ ] 已在目标设备/环境验证
- [ ] 不影响现有功能
- [ ] 已更新相关文档（如有需要）

## 备注
<!-- 其他需要说明的内容 -->

See merge request: CPF-KMP-CMP/kotlin!190

+++ kotlin-native/runtime/src/main/cpp/PointerBits.h
@@ -11,22 +11,22 @@
 #include "Common.h"
 
 template <typename T>
-ALWAYS_INLINE T* setPointerBits(T* ptr, unsigned bits) {
+ALWAYS_INLINE T* setPointerBits(T* ptr, uintptr_t bits) {
     return reinterpret_cast<T*>(reinterpret_cast<uintptr_t>(ptr) | bits);
 }
 
 template <typename T>
-ALWAYS_INLINE T* clearPointerBits(T* ptr, unsigned bits) {
-    return reinterpret_cast<T*>(reinterpret_cast<uintptr_t>(ptr) & ~static_cast<uintptr_t>(bits));
+ALWAYS_INLINE T* clearPointerBits(T* ptr, uintptr_t bits) {
+    return reinterpret_cast<T*>(reinterpret_cast<uintptr_t>(ptr) & ~bits);
 }
 
 template <typename T>
-ALWAYS_INLINE unsigned getPointerBits(T* ptr, unsigned bits) {
-    return reinterpret_cast<uintptr_t>(ptr) & static_cast<uintptr_t>(bits);
+ALWAYS_INLINE uintptr_t getPointerBits(T* ptr, uintptr_t bits) {
+    return reinterpret_cast<uintptr_t>(ptr) & bits;
 }
 
 template <typename T>
-ALWAYS_INLINE bool hasPointerBits(T* ptr, unsigned bits) {
+ALWAYS_INLINE bool hasPointerBits(T* ptr, uintptr_t bits) {
     return getPointerBits(ptr, bits) != 0;
 }
